get_epoch_ms is now long long so it doenst overflow

+++ weebware - cheat/weebware/shared_funcs.cpp
@@ -85,7 +85,7 @@ int get_config_index()
 }
 
 
-long get_epoch_ms() {
+long long get_epoch_ms() {
 	auto duration = std::chrono::system_clock::now().time_since_epoch();
 	return std::chrono::duration_cast<std::chrono::milliseconds>(duration).count();
 }
